ugc-service/etl-progress/pyproject.toml
[tool.poetry]
name = "Watching progress ETL service"
description = ""
version = "0.1.0"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
redis = "^4.6.0"
pydantic = "^1.10.7"
kafka-python = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
